refactor(qr): EventQRModal 445줄 본체를 useEventQRController 훅 + QRPanel 분리
EventQRModal(694줄)을 ①useEventQRController(상태/이펙트/데이터/파생플래그/핸들러 캡슐화)
②QRPanel(프레젠테이션: 3 서브컴포넌트+qrPanelContent 분기+JSX) ③EventQRModal(43줄 thin
shell)로 분할. public export(EventQRModal/EventQRModalProps) + Modal 래퍼 props 동일.

핵심 불변 보존(검증): lastGeneratedSignatureRef는 훅 내부에만(누수 0), 3개 이펙트·
generate/deactivate/isActive 캡슐화, 중복 regenerate 2곳을 onRegenerate로 일원화,
QRRefreshOverlay 애니메이션 verbatim. dark: 클래스 39=39 전수 보존.

검증: tsc 0 · eslint 0 · eventQRScope 4/4 · 원본 대비 이펙트/핸들러/modeColor/Modal
래퍼 byte-동일 diff 확인. (컴포넌트 테스트 부재 — 최종 적대검증 워크플로우로 재확인 예정)
